refactor(Edit Branch View): Clean up code
